requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Imbalanced Learning
imbalanced-learn>=0.11.0

# Gradient Boosting
xgboost>=2.0.0

# Deep Learning
tensorflow>=2.13.0

# Model Interpretability
shap>=0.42.0

# Utilities
joblib>=1.3.0
pyyaml>=6.0

# Development Tools
jupyter>=1.0.0
ipython>=8.12.0
